Start isolating dataclass logic in a module.
